feat: load all real images
used a script to download 368 plant images, backroundImageLoader loads all images in the folder, keys in hashmaps changed to scientificName, gardenEditorView shows first 10 images

+++ src/main/java/Plant.java
@@ -11,8 +11,8 @@ public class Plant implements Serializable {
 	private String color;
 	private double sizeLower;
 	private double  sizeUpper;
-	private int spreadRadiusLower;
-	private int spreadRadiusUpper;
+	private double spreadRadiusLower;
+	private double spreadRadiusUpper;
 	private Options options;
 	private double cost;
 	private int lepsSupported;
@@ -20,7 +20,7 @@ public class Plant implements Serializable {
 	private String genera;
 	
 	public Plant(String commonName, String scientificName, String genera, String family, String color, double sizeLower, double sizeUpper, 
-			int spreadRadiusLower, int spreadRadiusUpper, Options options, double cost, int lepsSupported, char type) {
+			double spreadRadiusLower, double spreadRadiusUpper, Options options, double cost, int lepsSupported, char type) {
 		this.commonName = commonName;
 		this.scientificName = scientificName;
 		this.genera = genera;
@@ -165,25 +165,25 @@ public class Plant implements Serializable {
 
 
 
-	public int getSpreadRadiusLower() {
+	public double getSpreadRadiusLower() {
 		return spreadRadiusLower;
 	}
 
 
 
-	public void setSpreadRadiusLower(int spreadRadiusLower) {
+	public void setSpreadRadiusLower(double spreadRadiusLower) {
 		this.spreadRadiusLower = spreadRadiusLower;
 	}
 
 
 
-	public int getSpreadRadiusUpper() {
+	public double getSpreadRadiusUpper() {
 		return spreadRadiusUpper;
 	}
 
 
 
-	public void setSpreadRadiusUpper(int spreadRadiusUpper) {
+	public void setSpreadRadiusUpper(double spreadRadiusUpper) {
 		this.spreadRadiusUpper = spreadRadiusUpper;
 	}
